Split PianoRoll into multiple files.
