+++ Sources/AudioKitUI/Controls/PianoRoll/PianoRollNoteView.swift
@@ -2,45 +2,6 @@
 
 import SwiftUI
 
-public struct PianoRollNote: Equatable, Identifiable {
-    public init(start: Double, length: Double, pitch: Int) {
-        self.start = start
-        self.length = length
-        self.pitch = pitch
-    }
-
-    public var id = UUID()
-
-    /// The start step.
-    var start: Double
-
-    /// How many steps long?
-    var length: Double
-
-    /// Abstract pitch, not MIDI notes.
-    var pitch: Int
-
-}
-
-public struct PianoRollModel: Equatable {
-    public init(notes: [PianoRollNote], length: Int, height: Int) {
-        self.notes = notes
-        self.length = length
-        self.height = height
-    }
-
-    /// The sequence being edited.
-    var notes:[PianoRollNote]
-
-    /// How many steps in the piano roll.
-    var length: Int
-
-    /// Maximum pitch.
-    var height: Int
-
-}
-
-
 struct PianoRollNoteView: View {
     @Binding var note: PianoRollNote
     var gridSize: CGSize
@@ -156,99 +117,5 @@ struct PianoRollNoteView: View {
                height: gridSize.height)
         .offset(noteOffset(note: note, dragOffset: offset))
 
-    }
-}
-
-struct PianoRollGrid: Shape {
-
-    var gridSize: CGSize
-    var length: Int
-    var height: Int
-
-    func path(in rect: CGRect) -> Path {
-
-        let size = rect.size
-        var path = Path()
-        for i in 0 ... length {
-            let x = CGFloat(i) * gridSize.width
-            path.move(to: CGPoint(x: x, y: 0))
-            path.addLine(to: CGPoint(x: x, y: size.height))
-        }
-
-        for i in 0 ... height {
-            let y = CGFloat(i) * gridSize.height
-            path.move(to: CGPoint(x: 0, y: y))
-            path.addLine(to: CGPoint(x: size.width, y: y))
-        }
-
-        return path
-    }
-}
-
-/// Touch-oriented piano roll.
-///
-/// Note: Requires macOS 12 / iOS 15 due to SwiftUI bug (crashes in SwiftUI when deleting notes).
-public struct PianoRoll: View {
-
-    @Binding var model: PianoRollModel
-    var gridSize = CGSize(width: 80, height: 40)
-    var noteColor = Color.accentColor
-
-    public init(model: Binding<PianoRollModel>, noteColor: Color = .accentColor) {
-        _model = model
-        self.noteColor = noteColor
-    }
-
-    let gridColor = Color(red: 15.0/255.0, green: 17.0/255.0, blue: 16.0/255.0)
-
-    public var body: some View {
-        ZStack(alignment: .topLeading) {
-            let dragGesture = DragGesture(minimumDistance: 0).onEnded({ value in
-                let location = value.location
-                let step = Double(Int(location.x / gridSize.width))
-                let pitch = model.height - Int(location.y / gridSize.height)
-                model.notes.append(PianoRollNote(start: step, length: 1, pitch: pitch))
-            })
-            PianoRollGrid(gridSize: gridSize, length: model.length, height: model.height)
-                .stroke(lineWidth: 0.5)
-                .foregroundColor(gridColor)
-                .contentShape(Rectangle())
-                .gesture(TapGesture().sequenced(before: dragGesture))
-            ForEach(model.notes) { note in
-                PianoRollNoteView(
-                    note: $model.notes[model.notes.firstIndex(of: note)!],
-                    gridSize: gridSize,
-                    color: noteColor,
-                    sequenceLength: model.length,
-                    sequenceHeight: model.height,
-                    isContinuous: true)
-                .onTapGesture {
-                    model.notes.removeAll(where: { $0 == note })
-                }
-            }
-        }.frame(width: CGFloat(model.length) * gridSize.width,
-                height: CGFloat(model.height) * gridSize.height)
-    }
-}
-
-public struct PianoRollTestView: View {
-
-    public init() { }
-
-    @State var model = PianoRollModel(notes: [
-        PianoRollNote(start: 1, length: 2, pitch: 3),
-        PianoRollNote(start: 5, length: 1, pitch: 4)
-    ], length: 128, height: 128)
-
-    public var body: some View {
-        ScrollView([.horizontal, .vertical], showsIndicators: true) {
-            PianoRoll(model: $model, noteColor: .cyan)
-        }.background(Color(white: 0.1))
-    }
-}
-
-struct PianoRoll_Previews: PreviewProvider {
-    static var previews: some View {
-        PianoRollTestView()
     }
 }
